core/baiss/requirements.txt
uvicorn
baisstools
fastapi
tiktoken
dotenv
httpx
pydantic_settings
pandas
pdfplumber
scikit-learn
matplotlib
huggingface_hub
transformers
duckdb
python-multipart
msoffcrypto-tool
python-docx
tabulate
nltk
openpyxl
websockets
FlashRank